Retrieve the hostname from the remote client
The cittad daemon gets the ip address of the remote host and transmits
this information to the remote_cittaclient when it execs it. The remote
client finds the corresponding host name, and notifies it to the server
with the INFO command. The server saves the host name of the user in
session->host if available, otherwise it saves the ip address. This is
what is displayed as host in the <w>ho and <p>rofile commands.