Improve <Enter> and line merge in internal editor
When hitting Enter on space between two words to break the line in two,
eliminate the leading space in the new line and trailing space above.
When merging two lines and an extra space is needed to keep words
separated, set the background color of the space to the default color
unless the background colors of the characters on the sides of the space
are the same.